requirements.txt
# Dependencies for automatic subtitle generator
openai-whisper==20231117
ffmpeg-python==0.2.0
keybert==0.8.4  # optional for keyword extraction
tqdm==4.66.1    # progress bars
rich==13.7.0    # enhanced terminal output
